# PERF: Faster comparisons of indexes when compared to self