@jbrockmendel 2020-10-13
that works for int dtype, but what about for dtypes that have NAs?