@twoertwein 2020-10-14
you could return `~isna()` for equal and `isna()` for not equal (if that is still faster)?