@jbrockmendel 2020-10-14
> you could return ~isna() for equal and isna() for not equal (if that is still faster)?

even then you'd need to watch out for pd.NA

I think just check for _can_hold_na.

Also for RangeIndex could compare the range objects instead of (i think) casting to ndarray