@topper-123 2020-10-14
I've made an update. I can't find the tests for index comparisons, they clearly need some tests for comparisons with indexes that contain nans. My first proposal obviously shouldn't have passed.

Anyonw know where they're located?