@jbrockmendel 2020-11-23
would this let us remove NumericIndex._cmp_method?