@topper-123 2021-01-01
This has already been fixed, e.g. the last example:

```python
>>> idx = rng.astype(object)
>>> %timeit idx == idx
5.97 ms ± 12.7 µs per loop  # when this PR was opened
53.9 µs ± 9.8 µs per loop  # master
```

So I'll just close this one.